fix(ngrx-ducks): avoid implicit any error in extractWiredAction
