+++ ngrx/ducks/src/lib/typings/class-with-action-annotations.ts
@@ -1,0 +1,5 @@
+export type ClassWithActionAnnotations<T> = {
+  [key: string]: T[keyof T] & {
+    wiredAction: { type: string; caseReducer: T[keyof T] };
+  };
+};
