+++ ngrx/ducks/src/lib/typings/index.ts
@@ -1,3 +1,4 @@
 export * from './action-that-may-have-a-payload';
+export * from './class-with-action-annotations';
 export * from './initial-state-annotated';
 export * from './reducer-function';
